backend/src/utils/requirements.txt
Pillow
requests
pymongo